+++ R/check.R
@@ -14,9 +14,7 @@
 #'  * 2: No clean-up and display build steps
 #' @importFrom rlang .data
 #' @export
-check = function(server,
-                 dir = ".",
-                 debug_level = 0:2) {
+check = function(server, dir = ".", debug_level = 0:2) {
   debug_level = audit.base::get_debug_level(force(debug_level))
 
   check_list = list()
@@ -28,12 +26,14 @@ check = function(server,
   check_list$sys_deps = check_sys_libs(debug_level)
   check_list$versions = check_versions(debug_level)
   cli::cli_h2("Starting checks")
-  r6_inits = audit.base::init_r6_checks(dir = dir,
-                                        file = "config-uat-psw.yml",
-                                        pkg_name = "audit.workbench")
+  r6_inits = audit.base::init_r6_checks(
+    dir = dir,
+    file = "config-uat-psw.yml",
+    pkg_name = "audit.workbench"
+  )
   lapply(r6_inits, function(r6) r6$check(debug_level = debug_level))
-  results = purrr::map_dfr(r6_inits, ~.x$get_log())
+  results = purrr::map_dfr(r6_inits, ~ .x$get_log())
   check_list$results = dplyr::arrange(results, .data$group, .data$short)
   cli::cli_h1("All checks complete")
-  return(invisible(check_list))
+  invisible(check_list)
 }